backend/expensetracker/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.52.2
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
backoff==2.2.1
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==45.0.4
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
Django==5.2
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==4.1
django-ratelimit==4.1.0
djangorestframework==3.16.0
dnspython==2.7.0
dotenv==0.9.9
dydantic==0.0.8
execnet==2.1.1
filelock==3.18.0
filetype==1.2.0
firebase-admin==6.9.0
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.1
fuzzywuzzy==0.18.0
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-api-python-client==2.173.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.1.1
google-crc32c==1.7.1
google-genai==1.22.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.73.1
gunicorn==21.2.0
h11==0.16.0
h2==4.2.0
hf-xet==1.1.5
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-anthropic==0.3.16
langchain-community==0.3.26
langchain-core==0.3.66
langchain-google-genai==2.1.5
langchain-mongodb==0.6.2
langchain-openai==0.3.26
langchain-text-splitters==0.3.8
langgraph==0.5.0
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.5.1
langgraph-sdk==0.1.70
langmem==0.0.27
langsmith==0.4.1
lark==1.2.2
markdown-it-py==3.0.0
marshmallow==3.26.1
mdurl==0.1.2
mongoengine==0.29.1
monotonic==1.6
msgpack==1.1.1
multidict==6.5.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.3.1
ollama==0.5.1
openai==1.92.2
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pillow==11.2.1
pluggy==1.6.0
portalocker==3.2.0
posthog==3.25.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.10.0
pydantic_core==2.33.2
pyfiglet==1.0.3
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.12.1
pyparsing==3.2.3
pytest==8.4.1
pytest-asyncio==1.0.0
pytest-repeat==0.9.4
pytest-rerunfailures==12.0
pytest-xdist==3.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9.1
safetensors==0.5.3
sentry-sdk==2.31.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
tabulate==0.9.0
tenacity==8.5.0
tiktoken==0.9.0
tokenizers==0.21.2
tqdm==4.67.1
transformers==4.52.4
trustcall==0.0.39
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.5.0
websockets==15.0.1
wheel==0.45.1
whitenoise==6.6.0
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
